Move _DictManager into h5store.py.
